self_attested_streamlit/requirements.txt

streamlit
pillow
PyPDF2
streamlit-cropper
